+++ commons-module/pom.xml
@@ -27,6 +27,10 @@
             <version>1.10.4</version>
             <scope>test</scope>
         </dependency>
+        <dependency>
+            <groupId>org.slf4j</groupId>
+            <artifactId>slf4j-simple</artifactId>
+        </dependency>
     </dependencies>
 
     <build>
@@ -48,5 +52,4 @@
             </plugin>
         </plugins>
     </reporting>
-
 </project>
