+++ pom.xml
@@ -19,23 +19,20 @@
 
     <properties>
         <compiler.version>1.8</compiler.version>
-
+        <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <allure.version>1.4.23</allure.version>
         <selenium.version>2.53.0</selenium.version>
         <htmlelements.version>1.15</htmlelements.version>
         <aspectj.version>1.8.5</aspectj.version>
     </properties>
 
-    <dependencies>
-        <dependency>
-            <groupId>junit</groupId>
-            <artifactId>junit</artifactId>
-            <version>4.12</version>
-        </dependency>
-    </dependencies>
-
     <dependencyManagement>
         <dependencies>
+            <dependency>
+                <groupId>org.slf4j</groupId>
+                <artifactId>slf4j-simple</artifactId>
+                <version>1.7.21</version>
+            </dependency>
             <dependency>
                 <groupId>ru.yandex.qatools.allure</groupId>
                 <artifactId>allure-junit-adaptor</artifactId>
@@ -65,6 +62,11 @@
                 <groupId>ru.yandex.qatools.htmlelements</groupId>
                 <artifactId>htmlelements-matchers</artifactId>
                 <version>1.12</version>
+            </dependency>
+            <dependency>
+                <groupId>junit</groupId>
+                <artifactId>junit</artifactId>
+                <version>4.12</version>
             </dependency>
         </dependencies>
     </dependencyManagement>
@@ -104,9 +106,30 @@
                         </dependency>
                     </dependencies>
                 </plugin>
+                <plugin>
+                    <groupId>org.apache.maven.plugins</groupId>
+                    <artifactId>maven-site-plugin</artifactId>
+                    <version>3.3</version>
+                </plugin>
+                <plugin>
+                    <groupId>org.apache.maven.plugins</groupId>
+                    <artifactId>maven-project-info-reports-plugin</artifactId>
+                    <version>2.9</version>
+                </plugin>
+                <!--Needed only to show reports locally. Run jetty:run and
+                    open localhost:8080 to show the report-->
+                <plugin>
+                    <groupId>org.eclipse.jetty</groupId>
+                    <artifactId>jetty-maven-plugin</artifactId>
+                    <version>9.2.10.v20150310</version>
+                    <configuration>
+                        <webAppSourceDirectory>${project.basedir}/commons-module/target/site/allure-maven-plugin
+                        </webAppSourceDirectory>
+                        <stopKey>stop</stopKey>
+                        <stopPort>1234</stopPort>
+                    </configuration>
+                </plugin>
             </plugins>
         </pluginManagement>
     </build>
-
-
 </project>
