+++ steps-module/pom.xml
@@ -41,5 +41,4 @@
             <artifactId>htmlelements-matchers</artifactId>
         </dependency>
     </dependencies>
-
 </project>
