+++ steps-module/src/main/java/ru/qatools/school/rules/WebDriverRule.java
@@ -1,6 +1,11 @@
 package ru.qatools.school.rules;
 
 import org.junit.rules.ExternalResource;
+import org.junit.rules.TestRule;
+import org.junit.runner.Description;
+import org.junit.runners.model.Statement;
+import org.openqa.selenium.By;
+import org.openqa.selenium.Keys;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.firefox.FirefoxDriver;
